Перенес всю логику с main.js в gallery.js
